Add back a VBlank wait loop to the reset vector
This partially undoes a change in
a129e119886e88d31fde8eb33d89560070dcc077.

The initialization code currently takes about 19K cycles to run, which
is less than the ~30K cycles or so on an NTSC NES where writes to the
PPU registers have no effect. (This can be emulated in Mesen using the
"Clear PPU $2000/1/5/6/7 throughout the first frame after reset/power
on" setting, though regional console differences don't seem to be
emulated properly.)

Because of that, the removal of the loops to wait for the PPU causes
NMI to not be enabled on an NTSC NES, meaning the game hangs during
boot once the game tries to wait for it to occur. Adding one of the
loops back allows the game to boot properly again.